Before enqueueing the job register it with the exact time it was supposed to run
Rufus calls the block with the current time instead of the time
the job is scheduled to run. If you have multiple hosts, it's possible
that Rufus will run the block at slightly different times which results
in a job getting ran multiple times.

This figures out when the job was scheduled to run instead of the time
rufus sends.

Co-Authored-By: Aaron Frase <afrase@users.noreply.github.com>