fix config check when sidekiq 6.5
we can't use key? directly and we can't use sidekiq_config.options.key?
because that will issue a warning, so we need to check if the value
returned by sidekiq_config[option] is nil, and assume that if it is, the
option is not present at the root level of the sidekiq config