Don't use depicated Redis commands
Sidekiq 7.1 now warns on depircated Redis commands. `zrangebyscore` is currently used by the gem which is depicated in Redis 6.2 which is the mimimum required Redis version for current Sidekiq.

Since we most likely are supporting the latest 7.x of Sidekiq, I reused the "GTE 7.0" check and redirected calls to `zrangebyscore` to use the Sidekiq Adaptor class.

The difference in commands is based off of the changes to Sidekiq when it swapped out the depricated call:
https://github.com/sidekiq/sidekiq/commit/1aca434e5cd830cbd0020f1f82ef3a3cb6b2d477#diff-25f829140be25a886134933b537d356dddb12234ca85319e50b381f9edb6b74dR686