Change cache-control to private
As mentioned in https://github.com/sidekiq/sidekiq/issues/5936 the public cache control could leak information when used with CloudFront or other CDNs. It was decided to change the cache setting in Sidekiq web's static assets to private.

This changes the gem's cache-control settings to match the recently released Sidekiq's.