Use correct folder structure for assets
The new world order `.register` method for adding pages in the Sidkeiq web UI expects a very specific folder structure of `{webdir}/assets/{tabname}/{assetdir}/{thefile}`.

The gem currently has the structure of `{webdir}/assets/{assetdir}/{thefile}`. Because of this, the latest Sidekiq was mounting the wrong folder paths for the static middleware.

- Moved the CSS file into the now expected folder structure.
- Updated the `<link>` tag in both the old and new implementation in the view to use the changed folder structure.
- Updated the old implementation of registering the page to use the changed folder structure for assets.