Remove check of sidekiq scheduler options that have moved under the  key
This check isn't necessary anymore for v6 — people migrating from v4 to v5 need to see this error, but from v5 to v6, it doesn't make sense to run this check since in theory users won't have the sidekiq scheduler configs at the root level anymore.

It could be that someone is migrating from v4 to v6, but then, it's on them to take this risk instead of upgrading 1 major version at time.

Closes https://github.com/sidekiq-scheduler/sidekiq-scheduler/issues/469